split active stream socket
